#13960 apply review suggestions
