+++ plugins/org.jkiss.dbeaver.core/src/org/jkiss/dbeaver/core/DesktopPreferencesInitializer.java
@@ -48,7 +48,7 @@ public class DesktopPreferencesInitializer extends AbstractPreferenceInitializer
         PrefUtils.setDefaultPreferenceValue(store, DBeaverPreferences.AGENT_ENABLED, true);
         PrefUtils.setDefaultPreferenceValue(store, DBeaverPreferences.AGENT_LONG_OPERATION_NOTIFY, RuntimeUtils.isWindows());
         PrefUtils.setDefaultPreferenceValue(store, DBeaverPreferences.AGENT_LONG_OPERATION_TIMEOUT, 30);
-        PrefUtils.setDefaultPreferenceValue(store, DBeaverPreferences.CLIENT_TIMEZONE, "DEFAULT");
+        PrefUtils.setDefaultPreferenceValue(store, DBeaverPreferences.CLIENT_TIMEZONE, null);
         PrefUtils.setDefaultPreferenceValue(store, DBeaverPreferences.SECURITY_USE_BOUNCY_CASTLE, true);
 
         PrefUtils.setDefaultPreferenceValue(store, DBeaverPreferences.NAVIGATOR_EDITOR_FULL_NAME, false);
