+++ plugins/org.jkiss.dbeaver.core/src/org/jkiss/dbeaver/registry/TimezoneRegistry.java
@@ -20,40 +20,68 @@ import org.jkiss.code.NotNull;
 import org.jkiss.dbeaver.DBeaverPreferences;
 import org.jkiss.dbeaver.model.preferences.DBPPreferenceStore;
 import org.jkiss.dbeaver.runtime.DBWorkbench;
-import org.jkiss.dbeaver.utils.TimezoneUtils;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.TimeZone;
+import java.util.concurrent.TimeUnit;
 
 public class TimezoneRegistry {
 
-    private static DBPPreferenceStore preferenceStore;
     public static final String DEFAULT_VALUE = "N/A";
     private TimezoneRegistry() {
     }
 
     public static void setUsedTime(@NotNull String text) {
-        initStore();
+        DBPPreferenceStore preferenceStore = DBWorkbench.getPlatform().getPreferenceStore();
         if (!text.equals(DEFAULT_VALUE)) {
-            preferenceStore.setValue(DBeaverPreferences.CLIENT_TIMEZONE, TimezoneUtils.extractTimezoneId(text));
-            TimeZone.setDefault(TimeZone.getTimeZone(TimezoneUtils.extractTimezoneId(text)));
+            preferenceStore.setValue(DBeaverPreferences.CLIENT_TIMEZONE, extractTimezoneId(text));
+            TimeZone.setDefault(TimeZone.getTimeZone(extractTimezoneId(text)));
         } else {
-            preferenceStore.setValue(DBeaverPreferences.CLIENT_TIMEZONE, text);
+            preferenceStore.setToDefault(DBeaverPreferences.CLIENT_TIMEZONE);
             TimeZone.setDefault(null);
         }
     }
 
     public static void overrideTimezone() {
-        initStore();
+        DBPPreferenceStore preferenceStore = DBWorkbench.getPlatform().getPreferenceStore();
         final String timezone = preferenceStore.getString(DBeaverPreferences.CLIENT_TIMEZONE);
         if (timezone != null && !timezone.equals(DEFAULT_VALUE)) {
             TimeZone.setDefault(TimeZone.getTimeZone(timezone));
         }
     }
 
-    private static void initStore() {
-        if (preferenceStore == null) {
-            preferenceStore = DBWorkbench.getPlatform().getPreferenceStore();
+    @NotNull
+    public static List<String> getTimezoneNames() {
+        List<String> list = new ArrayList<>();
+        for (String s : TimeZone.getAvailableIDs()) {
+            String addGMTTime = addGMTTime(s);
+            list.add(addGMTTime);
         }
+        return list;
     }
+
+    @NotNull
+    public static String addGMTTime(@NotNull String availableID) {
+        String result;
+        long hours = TimeUnit.MILLISECONDS.toHours(TimeZone.getTimeZone(availableID).getRawOffset());
+        long minutes = TimeUnit.MILLISECONDS.toMinutes(TimeZone.getTimeZone(availableID).getRawOffset())
+            - TimeUnit.HOURS.toMinutes(hours);
+        minutes = Math.abs(minutes);
+        String hoursWithSymbol;
+        if (hours < 0) {
+            hoursWithSymbol = Long.toString(hours);
+        } else {
+            hoursWithSymbol = '+' +Long.toString(hours);
+        }
+        result = String.format("%s (UTC%s:%02d)", availableID, hoursWithSymbol, minutes);
+        return result;
+    }
+
+
+    public static String extractTimezoneId(@NotNull String timezone) {
+        return timezone.split(" ")[0];
+    }
+
 
 }
